index.api 충돌 수정 및 소식 리스트 조회 변경
