+++ src/api/index.api.ts
@@ -1,77 +1,26 @@
-import { Api, Parameter } from '@Interface/api.interface';
+import { Api, IParameter, Parameter } from '@Interface/api.interface';
 import {
-  EndPointType,
-  QueryOrderByOptions,
   QueryWhereOptions,
+  QueryOrderByOptions,
+  EndPointType,
 } from '@Type/api.type';
+import firebase from 'firebase/compat/app';
 import {
   addDoc,
-  setDoc,
   collection,
   doc,
   DocumentData,
+  FieldPath,
   getDoc,
   getDocs,
-  QuerySnapshot,
-  Timestamp,
-  WhereFilterOp,
   QueryDocumentSnapshot,
-  FieldPath,
-  OrderByDirection,
+  QuerySnapshot,
+  setDoc,
+  Timestamp,
 } from 'firebase/firestore';
-import { isEmpty } from 'lodash';
 import { db } from './firebase';
-import firebase from 'firebase/compat/app';
 
-/* endpoint type */
-export type EndPointType = 'news' | 'work' | 'profile' | 'members';
-
-/* query parameter type */
-export type QueryType = 'where' | 'orderby';
-
-export type QueryWhereOptions = {
-  fieldPath: string | FieldPath;
-  opStr: WhereFilterOp;
-  value: any;
-};
-export type QueryOrderByOptions = {
-  fieldPath: string | FieldPath;
-  directionStr?: OrderByDirection;
-  limit?: number;
-};
-
-export type TQuery =
-  | ({
-      queryType: 'where';
-    } & QueryWhereOptions)
-  | ({ queryType: 'orderby' } & QueryOrderByOptions);
-
-/**
- * @param endPoint (require) firebase collection name
- * @param param    (require) id: 'Document ID' firebase sequence value + etc...
- * @param searchFields (optional) search specific field: string[]
- */
-export interface Parameter<U = any> {
-  endPoint: EndPointType;
-  param: U;
-  searchField?: string[];
-}
-
-export interface IParameter {
-  endPoint: EndPointType;
-  queries: TQuery[];
-  searchFields?: string[];
-}
-
-declare interface Api {
-  <Request extends Parameter<Request>, Response>(
-    param: Parameter,
-  ): Promise<Response>;
-  <Request extends Parameter<Request>, Response>(
-    param: Parameter,
-  ): Promise<Response>;
-}
-
+// TODO: 삭제해야함
 export const GetData: Api = async ({ endPoint, param }) => {
   try {
     const docRef = doc(db, endPoint, param.id);
@@ -98,29 +47,6 @@ export const GetDataList = async <U>({
   }
 };
 
-export const GetDataListQuery: Api = async <
-  U extends { conditions: QueryWhereOptions[]; orderBy?: QueryOrderByOptions },
->({
-  endPoint,
-  param,
-  searchField,
-}: Parameter<U>) => {
-  try {
-    const ref = db.collection(endPoint);
-    const resultData = (await getMultipleWhereQueries(
-      ref,
-      param.conditions,
-    ).get()) as QuerySnapshot;
-
-    return resultData.empty
-      ? []
-      : resultData.docs.map(doc => getData(doc, searchField));
-  } catch (e) {
-    console.error(e);
-    throw e;
-  }
-};
-
 export const GetDataListQueryOrderBy: Api = async <
   U extends QueryOrderByOptions,
 >({
@@ -138,7 +64,7 @@ export const GetDataListQueryOrderBy: Api = async <
           orderByRef = orderByRef.limit(param.limit);
         }
 
-        orderByRef.onSnapshot(docs => {
+        orderByRef.onSnapshot((docs: { docs: any[] }) => {
           docs.docs.forEach(doc => {
             resultData.push({
               id: doc.id,
@@ -158,6 +84,7 @@ export const GetDataListQueryOrderBy: Api = async <
     throw e;
   }
 };
+
 /*
   U: interface type of response array.
 */
